月卡分账

+++ src/main/java/com/xiliulou/electricity/mapper/ElectricityMemberCardMapper.java
@@ -12,7 +12,7 @@ public interface ElectricityMemberCardMapper extends BaseMapper<ElectricityMembe
 
     List<ElectricityMemberCard> queryList(@Param("offset") Long offset, @Param("size") Long size, @Param("status") Integer status, @Param("type") Integer type,@Param("tenantId") Integer tenantId);
 
-	List<ElectricityMemberCard> queryUserList(@Param("offset") Long offset, @Param("size") Long size, @Param("id") Integer id);
+	List<ElectricityMemberCard> queryUserList(@Param("offset") Long offset, @Param("size") Long size, @Param("id") Long id);
 
 	Integer queryCount(@Param("status") Integer status, @Param("type") Integer type,@Param("tenantId") Integer tenantId);
